requirements.txt
fastapi
gunicorn
uvicorn
sqlalchemy
python-dotenv
ezgmail
psycopg2
google-genai